+++ CHANGES.md
@@ -64,7 +64,7 @@ versioning](https://semver.org/) and all releases are available on
   Kindergeld is not set to 0 if Kinderfreibetrag is beneficial
   ({ghuser}`LauraGergeleit`, {ghuser}`ChristianZimpelmann`).
 
-- {gh}`463` Fix bug in calculation of income for Kinderzuschlag.
+- {gh}`461` Fix bug in calculation of income for Kinderzuschlag.
   ({ghuser}`ChristianZimpelmann`).
 
 ## 0.5.1 — 2022-04-21